# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nisztor-Bukovina Platform API
  version: 0.1.0
  description: Contract-first REST API for the Nisztor-Bukovina Platform.
servers:
  - url: /api
paths:
...
  /booking-management/{token}:
    get:
      operationId: getBookingManagementSummary
      summary: Get a booking-request summary through its management token
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/BookingManagementToken'
        - $ref: '#/components/parameters/Language'
      responses:
        '200':
          description: Token-scoped, non-editable booking-request summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingManagementSummary'
        '404':
          $ref: '#/components/responses/BookingManagementLinkInvalid'
  /booking-management/{token}/cancellation:
    post:
      operationId: cancelBookingThroughManagementLink
      summary: Cancel a cancellable booking request through its management token
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/BookingManagementToken'
      responses:
        '204':
          description: The booking request was cancelled and the token was revoked.
        '404':
          $ref: '#/components/responses/BookingManagementLinkInvalid'
        '409':
          description: The valid token belongs to a booking that cannot be self-service cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingManagementError'
...
  /admin/guesthouses/content:
    get:
      operationId: listAdminGuesthouseContent
      summary: List editable content for every guesthouse
      description: Returns active and inactive guesthouses with raw HU, RO, and EN translations. Missing translations are returned as empty editable records without applying public fallback.
      tags: [Administration]
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: Guesthouses in display order with all editable translations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminGuesthouseContent'
        '401':
          description: Administrator authentication is missing or invalid.
        '403':
          description: The authenticated account is not an administrator.
  /admin/guesthouses/{guesthouseId}/translations/{language}:
    put:
      operationId: updateAdminGuesthouseTranslation
      summary: Create or replace one guesthouse translation
      description: Uses a per-translation version for optimistic concurrency. A missing translation is created when version is null.
      tags: [Administration]
      security:
        - AdminBearerAuth: []
      parameters:
        - name: guesthouseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: language
          in: path
          required: true
          schema:
            type: string
            enum: [hu, ro, en]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminGuesthouseTranslationUpdate'
      responses:
        '200':
          description: Saved translation with its current version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGuesthouseTranslation'
        '400':
          description: Unsupported language, malformed input, or field validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGuesthouseContentError'
        '401':
          description: Administrator authentication is missing or invalid.
        '403':
          description: The authenticated account is not an administrator.
        '404':
          description: No guesthouse exists with the supplied identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGuesthouseContentError'
        '409':
          description: The translation was changed after the administrator loaded it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGuesthouseContentError'
  /admin/guesthouses/{guesthouseId}/pricing:
    get:
      operationId: getAdminGuesthousePricing
      summary: Get the current editable prices of one guesthouse
      tags: [Administration]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuesthouseId'
      responses:
        '200':
          description: Current guesthouse prices, surcharges, and discounts.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminGuesthousePricing' }
        '401': { description: Administrator authentication is missing or invalid. }
        '403': { description: The authenticated account is not an administrator. }
        '404':
          description: No guesthouse or active price configuration exists with the supplied identifier.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminGuesthousePricingError' }
    put:
      operationId: updateAdminGuesthousePricing
      summary: Replace the current numeric prices of one guesthouse
      tags: [Administration]
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GuesthouseId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminGuesthousePricingUpdate' }
      responses:
        '200':
          description: Updated current guesthouse prices.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminGuesthousePricing' }
        '400':
          description: Invalid, missing, duplicate, negative, or unknown price values.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminGuesthousePricingError' }
        '401': { description: Administrator authentication is missing or invalid. }
        '403': { description: The authenticated account is not an administrator. }
        '404':
          description: No guesthouse or active price configuration exists with the supplied identifier.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminGuesthousePricingError' }
...
  /tourism/attractions:
    get:
      operationId: listPublicAttractions
      summary: List active attractions that have the requested translation
      tags: [Tourism]
      parameters:
        - { name: lang, in: query, schema: { $ref: '#/components/schemas/Language' } }
      responses:
        '200':
          description: Active attractions.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AttractionPublic' } }
  /tourism/collections:
    get:
      operationId: listPublicTourismCollections
      summary: List active editorial tourism collections that have the requested translation
      tags: [Tourism]
      parameters:
        - { name: lang, in: query, schema: { $ref: '#/components/schemas/Language' } }
      responses:
        '200':
          description: Active tourism collections.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TourismCollection' } }
  /tourism/attractions/{slug}:
    get:
      operationId: getPublicAttraction
      summary: Get one active attraction
      tags: [Tourism]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: lang, in: query, schema: { $ref: '#/components/schemas/Language' } }
      responses:
        '200':
          description: Attraction detail.
          content: { application/json: { schema: { $ref: '#/components/schemas/AttractionPublic' } } }
        '404': { description: Attraction not found or inactive. }
...
  /admin/tourism/attractions:
    get:
      operationId: listAdminAttractions
      tags: [Administration, Tourism]
      security: [{ AdminBearerAuth: [] }]
      responses:
        '200':
          description: All attractions including inactive records.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AdminAttraction' } }
        '401': { description: Administrator authentication is missing or invalid. }
    post:
      operationId: createAdminAttraction
      tags: [Administration, Tourism]
      security: [{ AdminBearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AttractionUpdate' } } }
      responses:
        '201':
          description: Attraction created.
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminAttraction' } } }
        '400': { description: Attraction validation failed. }
        '401': { description: Administrator authentication is missing or invalid. }
        '409': { description: Slug already exists. }
  /admin/tourism/attractions/{id}:
    put:
      operationId: updateAdminAttraction
      tags: [Administration, Tourism]
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AttractionUpdate' } } }
      responses:
        '200':
          description: Attraction updated.
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminAttraction' } } }
        '400': { description: Attraction validation failed. }
        '401': { description: Administrator authentication is missing or invalid. }
        '404': { description: Attraction not found. }
        '409': { description: Slug already exists. }
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    BookingManagementLinkInvalid:
      description: The management link is invalid, expired, revoked, or already used.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookingManagementError'
  parameters:
    GuesthouseId:
      name: guesthouseId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    BookingManagementToken:
      name: token
      in: path
      required: true
      description: Random URL-safe management token. Its raw value is never persisted.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{43}$'
    Language:
      name: lang
      in: query
      required: false
      description: Requested content language. Missing translations fall back to Hungarian.
      schema:
        type: string
        enum: [hu, ro, en]
        default: hu
  schemas:
...
    BookingManagementSummary:
      type: object
      required: [reference, status, guesthouse, stay, guests, services, rooms, price, cancellationAllowed]
      properties:
        reference: { type: string }
        status: { $ref: '#/components/schemas/BookingStatus' }
        guesthouse:
          type: object
          required: [name, contacts]
          properties:
            name: { type: string }
            contacts:
              type: array
              items:
                type: object
                required: [type, value, preferred]
                properties:
                  type: { type: string, enum: [PHONE, EMAIL] }
                  value: { type: string }
                  label: { type: [string, 'null'] }
                  preferred: { type: boolean }
        stay:
          type: object
          required: [checkInDate, checkOutDate, nights]
          properties:
            checkInDate: { type: string, format: date }
            checkOutDate: { type: string, format: date }
            nights: { type: integer, format: int64, minimum: 1 }
        guests:
          type: object
          required: [adults, childrenAge3to10, childrenAge0to3]
          properties:
            adults: { type: integer, minimum: 0 }
            childrenAge3to10: { type: integer, minimum: 0 }
            childrenAge0to3: { type: integer, minimum: 0 }
        services:
          type: object
          required: [breakfastParticipants, dinnerParticipants]
          properties:
            breakfastParticipants: { type: integer, minimum: 0 }
            dinnerParticipants: { type: integer, minimum: 0 }
        rooms:
          type: array
          items:
            type: object
            required: [name, quantity]
            properties:
              name: { type: string }
              quantity: { type: integer, minimum: 1 }
        price:
          type: object
          required: [accommodationTotal, breakfastTotal, dinnerTotal, totalPayable, currency]
          properties:
            accommodationTotal: { type: number, minimum: 0 }
            breakfastTotal: { type: number, minimum: 0 }
            dinnerTotal: { type: number, minimum: 0 }
            totalPayable: { type: number, minimum: 0 }
            currency: { type: string, enum: [RON] }
        cancellationAllowed: { type: boolean }
    BookingManagementError:
      type: object
      required: [code]
      properties:
        code:
          type: string
          enum: [BOOKING_MANAGEMENT_LINK_INVALID, BOOKING_CANCELLATION_NOT_ALLOWED]
...
    BookingStatus:
      type: string
      enum: [RECEIVED, UNDER_REVIEW, CONFIRMED, REJECTED, CANCELLED]
...
    AdminGuesthouseContent:
      type: object
      additionalProperties: false
      required: [id, slug, active, translations]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        active: { type: boolean }
        translations:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: '#/components/schemas/AdminGuesthouseTranslation'
...
    AdminGuesthouseTranslation:
      type: object
      additionalProperties: false
      required:
        - language
        - version
        - name
        - shortDescription
        - description
        - roomDescription
        - storyEyebrow
        - storyTitle
        - diningEyebrow
        - diningTitle
        - diningDescription
        - amenitiesTitle
        - roomTypesTitle
        - pricingTitle
        - historyEyebrow
        - historyTitle
        - historyText
        - galleryTitle
        - galleryHint
      properties:
        language: { type: string, enum: [hu, ro, en] }
        version:
          type: [integer, 'null']
          format: int64
          minimum: 0
          description: Null only when the language-specific translation does not exist yet.
        name: { type: string, maxLength: 160 }
        shortDescription: { type: string, maxLength: 500 }
        description: { type: string, maxLength: 5000 }
        roomDescription: { type: string, maxLength: 3000 }
        storyEyebrow: { type: string, maxLength: 240 }
        storyTitle: { type: string, maxLength: 240 }
        diningEyebrow: { type: string, maxLength: 240 }
        diningTitle: { type: string, maxLength: 240 }
        diningDescription: { type: string, maxLength: 1000 }
        amenitiesTitle: { type: string, maxLength: 240 }
        roomTypesTitle: { type: string, maxLength: 240 }
        pricingTitle: { type: string, maxLength: 240 }
        historyEyebrow: { type: string, maxLength: 240 }
        historyTitle: { type: string, maxLength: 240 }
        historyText: { type: string, maxLength: 5000 }
        galleryTitle: { type: string, maxLength: 240 }
        galleryHint: { type: string, maxLength: 500 }
    AdminGuesthouseTranslationUpdate:
      type: object
      additionalProperties: false
      required:
        - version
        - name
        - shortDescription
        - description
        - roomDescription
        - storyEyebrow
        - storyTitle
        - diningEyebrow
        - diningTitle
        - diningDescription
        - amenitiesTitle
        - roomTypesTitle
        - pricingTitle
        - historyEyebrow
        - historyTitle
        - historyText
        - galleryTitle
        - galleryHint
      properties:
        version:
          type: [integer, 'null']
          format: int64
          minimum: 0
        name: { type: string, minLength: 1, maxLength: 160 }
        shortDescription: { type: string, minLength: 1, maxLength: 500 }
        description: { type: string, minLength: 1, maxLength: 5000 }
        roomDescription: { type: string, minLength: 1, maxLength: 3000 }
        storyEyebrow: { type: string, minLength: 1, maxLength: 240 }
        storyTitle: { type: string, minLength: 1, maxLength: 240 }
        diningEyebrow: { type: string, minLength: 1, maxLength: 240 }
        diningTitle: { type: string, minLength: 1, maxLength: 240 }
        diningDescription: { type: string, minLength: 1, maxLength: 1000 }
        amenitiesTitle: { type: string, minLength: 1, maxLength: 240 }
        roomTypesTitle: { type: string, minLength: 1, maxLength: 240 }
        pricingTitle: { type: string, minLength: 1, maxLength: 240 }
        historyEyebrow: { type: string, minLength: 1, maxLength: 240 }
        historyTitle: { type: string, minLength: 1, maxLength: 240 }
        historyText: { type: string, minLength: 1, maxLength: 5000 }
        galleryTitle: { type: string, minLength: 1, maxLength: 240 }
        galleryHint: { type: string, minLength: 1, maxLength: 500 }
    AdminGuesthouseContentError:
      type: object
      additionalProperties: false
      required: [code, fieldErrors]
      properties:
        code:
          type: string
          enum:
            - ADMIN_GUESTHOUSE_NOT_FOUND
            - UNSUPPORTED_CONTENT_LANGUAGE
            - ADMIN_CONTENT_VALIDATION_FAILED
            - INVALID_ADMIN_CONTENT_REQUEST
            - ADMIN_CONTENT_VERSION_CONFLICT
        fieldErrors:
          type: object
          additionalProperties:
            type: string
            enum: [REQUIRED, TOO_LONG]
        currentContent:
          oneOf:
            - $ref: '#/components/schemas/AdminGuesthouseTranslation'
            - type: 'null'
    AdminGuesthousePricing:
      type: object
      additionalProperties: false
      required: [guesthouseId, currency, items, surcharges, discounts]
      properties:
        guesthouseId: { type: string, format: uuid }
        currency: { type: string, enum: [RON] }
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminPriceItem' }
        surcharges:
          type: array
          items: { $ref: '#/components/schemas/AdminPricingAdjustment' }
        discounts:
          type: array
          items: { $ref: '#/components/schemas/AdminPricingAdjustment' }
    AdminPriceItem:
      type: object
      additionalProperties: false
      required: [code, label, amount, unit]
      properties:
        code: { type: string }
        label: { type: string }
        amount: { type: number, minimum: 0 }
        unit: { type: string, enum: [person_night, room_night, person, day] }
    AdminPricingAdjustment:
      type: object
      additionalProperties: false
      required: [code, label, percentage]
      properties:
        code: { type: string }
        label: { type: string }
        percentage: { type: number, minimum: 0, maximum: 100 }
    AdminGuesthousePricingUpdate:
      type: object
      additionalProperties: false
      required: [items, surcharges, discounts]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [code, amount]
            properties:
              code: { type: string, minLength: 1 }
              amount: { type: number, minimum: 0 }
        surcharges:
          type: array
          items: { $ref: '#/components/schemas/AdminPricingAdjustmentUpdate' }
        discounts:
          type: array
          items: { $ref: '#/components/schemas/AdminPricingAdjustmentUpdate' }
    AdminPricingAdjustmentUpdate:
      type: object
      additionalProperties: false
      required: [code, percentage]
      properties:
        code: { type: string, minLength: 1 }
        percentage: { type: number, minimum: 0, maximum: 100 }
    AdminGuesthousePricingError:
      type: object
      additionalProperties: false
      required: [code, fieldErrors]
      properties:
        code:
          type: string
          enum:
            - ADMIN_PRICING_GUESTHOUSE_NOT_FOUND
            - ADMIN_PRICING_NOT_FOUND
            - ADMIN_PRICING_VALIDATION_FAILED
            - INVALID_ADMIN_PRICING_REQUEST
        fieldErrors:
          type: object
          additionalProperties: { type: string }
...
    AttractionTranslation:
      type: object
      required: [language, name, shortDescription, detailedDescription]
      properties:
        language: { $ref: '#/components/schemas/Language' }
        name: { type: string }
        shortDescription: { type: string }
        detailedDescription: { type: string, description: Rich-text-ready content stored as safe text in the MVP. }
        admissionInformation: { type: [string, 'null'] }
        practicalInformation: { type: [string, 'null'] }
    TourismCollection:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
        shortDescription: { type: [string, 'null'] }
    AttractionUpdate:
      type: object
      required: [slug, latitude, longitude, googleMapsUrl, active, translations, collectionSlugs]
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        googleMapsUrl: { type: string, format: uri }
        active: { type: boolean }
        translations: { type: array, minItems: 1, items: { $ref: '#/components/schemas/AttractionTranslation' } }
        collectionSlugs: { type: array, uniqueItems: true, items: { type: string } }
    AdminAttraction:
      allOf:
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
            distanceCalculation:
              oneOf:
                - $ref: '#/components/schemas/DrivingDistanceCalculation'
                - type: 'null'
        - { $ref: '#/components/schemas/AttractionUpdate' }
    DrivingDistanceCalculation:
      type: object
      description: Returned only after an admin creation or coordinate update has completed its Google Routes recalculation.
      required: [total, successful, failed]
      properties:
        total: { type: integer, minimum: 0 }
        successful: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
    AttractionPublic:
      type: object
      required: [slug, name, shortDescription, detailedDescription, latitude, longitude, googleMapsUrl, collectionSlugs]
      properties:
        slug: { type: string }
        name: { type: string }
        shortDescription: { type: string }
        detailedDescription: { type: string }
        admissionInformation: { type: [string, 'null'] }
        practicalInformation: { type: [string, 'null'] }
        latitude: { type: number }
        longitude: { type: number }
        googleMapsUrl: { type: string, format: uri }
        collectionSlugs: { type: array, items: { type: string } }
...